+++ src/commands/Moderation/lock.ts
@@ -5,8 +5,8 @@ import { Bot } from '../..';
 import { sendError } from '../../utils/messages';
 import { durationToString, stringToDuration, stringToChannel } from '../../utils/parsers';
 import { Durations } from '../../utils/time';
-import { GuildObject } from '../../database/schemas';
 import { GuildWrapper } from '../../database/wrappers/guildWrapper';
+import { GuildObject } from '../../database/schemas/main/guild';
 
 /**
  * returns the IDs that it has to overwrite and in what state they were
